+++ test/sql/pgsql.sql
@@ -33,6 +33,8 @@ CREATE TABLE venues (
 	state char(2),
 	address varchar(50),
 	phone varchar(10) default NULL,
+	is_available boolean DEFAULT true,
+	is_retired boolean DEFAULT false,
 	UNIQUE(name,address)
 );